Restructured project for Maven
